LayoutTests/accessibility/mac/selection-element-tabbing-to-link-expected.txt

one two
This tests that tabbing between links includes a relevant accessibilityObject in the userInfo when on the boundary between an ignored accessibilityObject and an unignored accessibilityObject.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS addedNotification is true
PASS ancestorWithRole(changeElementOne, "AXRole: AXLink").isEqual(linkTwo) is true
PASS ancestorWithRole(changeElementTwo, "AXRole: AXLink").isEqual(linkOne) is true
PASS successfullyParsed is true

TEST COMPLETE
